pyproject.toml
[project]
name = "PyDFLT"
version = "0.1.0"
description = "PyDFLT"
authors = [
    { name = "Noah Schutte", email = "N.J.Schutte@tudelft.nl"},
    { name = "Kim van den Houten", email = "K.C.vandenhouten@tudelft.nl"},
    { name = "Grigorii Veviurko", email = "G.Veviurko@tudelft.nl"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.7.1",
    "cvxpy>=1.4.1",
    "diffcp>=1.0.18",
    "cvxpylayers>=0.1.6",
    "numpy>=1.26.4",
    "ortools>=9.12.4544",
    "gurobipy>=11.0.3",
    "pyepo~=0.3.9",
    "pyyaml>=6.0.2",  # PyYAML is same package
    "tomli>=2.0.1",
    "wandb>=0.19.11",
    "pandas>=2.2.3",
    "matplotlib>=3.6.3",
    "optuna>=4.2.1",
    "optuna-dashboard>=0.19.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "joblib>=1.4.2",
    "ipywidgets>=8.1.7"
]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "nbmake>=1.5.5",
    "codecov>=2.1.13",
    "pre-commit>=3.8.0",
    "ruff>=0.12.4",
]

docs = [
    "sphinx>=7.4.7",
    "nbsphinx>=0.9.5",
    "numpydoc>=1.8.0",
    "sphinx-immaterial>=0.12.2",
]

examples = [
    "jupyter>=1.0.0",
]

[tool.uv]
required-version = "0.6.16"
default-groups = ["dev"]

[tool.black]
line-length = 160

[tool.ruff]
line-length = 160
extend-include = ["*.ipynb"]


[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "I",  # isort
    "W",  # pycodestyle warnings,
    "F",  # pyflakes,
    "C",  # flake8-comprehensions,
    "B",  # flake8-bugbear
]
ignore = [
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-third-party = ["numpy", "wandb"]  # can cause issues with pre-commit otherwise

[tool.pytest.ini_options]
addopts = "--cov=. --cov-report=xml --nbmake"
testpaths = [
    "tests",
    "examples",
]


[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"
